float64 선언되어 불필요하게 진행되는 업스케일링 제거, raw2temp를 사용하여 prob 수준에서 처리되는 부하 감소

+++ thermal_monitoring/analysis/roi.py
@@ -103,7 +103,7 @@ def extract_roi_from_npy(npy_path: str, config: Optional[RoiConfig] = None) -> R
     if config is None:
         config = load_roi_config()
 
-    thermal = np.load(npy_path).astype(np.float64)
+    thermal = np.load(npy_path).astype(np.float32)
 
     if thermal.ndim != 2:
         raise ValueError(f"Expected 2D thermal array, got shape {thermal.shape}")
